feat(backend): implement epic 2 - orders and inventory management

+++ app/models/requests.py
@@ -1,5 +1,5 @@
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
-from sqlalchemy import Enum as SQLEnum, ForeignKey
+from sqlalchemy import Enum as SQLEnum, ForeignKey, Numeric
 from sqlalchemy import Column, String, Boolean, Date,  Integer, DateTime
 from sqlalchemy.orm import relationship
 from uuid import UUID, uuid4
@@ -29,7 +29,7 @@ class PurchaseRequestItem(Base):
     id = Column(PG_UUID(as_uuid=True), primary_key=True, default=uuid4)
     request_id = Column(PG_UUID(as_uuid=True), ForeignKey("purchase_requests.id"), nullable=False, index=True)
     product_id = Column(PG_UUID(as_uuid=True), ForeignKey("products.id"), nullable=False, index=True)
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Numeric(10,2), nullable=False)
 
     # Relationships
     request = relationship("PurchaseRequest", back_populates="items")
